2017/11/9 update new Non-OS and Non-OS_Keil

+++ SDK/Non-OS/SIC/Src/sic.c
@@ -33,7 +33,7 @@
 #include <string.h>
 
 #include "wblib.h"
-#include "w55fa92_sic.h"
+#include "w55fa95_sic.h"
 #include "fmi.h"
 
 /*-----------------------------------------------------------------------------------*/
@@ -41,7 +41,7 @@
 /*   sicOpen                                                                         */
 /*                                                                                   */
 /* Parameters:                                                                       */
-/*   None                                                                            */
+/*   None				                                                             */
 /*                                                                                   */
 /* Returns:                                                                          */
 /*   None.                                                                           */
@@ -53,14 +53,14 @@
 /*   This function reset the DMAC and SIC interface and enable interrupt.            */
 /*                                                                                   */
 /*-----------------------------------------------------------------------------------*/
-static short _fmi_init_flag = FALSE;
+static int _fmi_init_flag = FALSE;
 void sicOpen(void)
 {
-    if (!_fmi_init_flag)
-    {
-        fmiInitDevice();
-        _fmi_init_flag = TRUE;
-    }
+	if (!_fmi_init_flag)
+	{
+		fmiInitDevice();
+		_fmi_init_flag = TRUE;
+	}
 }
 
 /*-----------------------------------------------------------------------------------*/
@@ -68,7 +68,7 @@ void sicOpen(void)
 /*   sicClose                                                                         */
 /*                                                                                   */
 /* Parameters:                                                                       */
-/*   None                                                                            */
+/*   None				                                                             */
 /*                                                                                   */
 /* Returns:                                                                          */
 /*   None.                                                                           */
@@ -82,9 +82,9 @@ void sicOpen(void)
 /*-----------------------------------------------------------------------------------*/
 void sicClose(void)
 {
-    outpw(REG_AHBCLK, inpw(REG_AHBCLK) & (~(SD_CKE | SIC_CKE)));
-    sysDisableInterrupt(IRQ_SIC);
-    _fmi_init_flag = FALSE;
+	outpw(REG_AHBCLK, inpw(REG_AHBCLK) & (~(SD_CKE | SIC_CKE)));
+	sysDisableInterrupt(IRQ_SIC);
+	_fmi_init_flag = FALSE;
 }
 
 /*-----------------------------------------------------------------------------------*/
@@ -93,9 +93,9 @@ void sicClose(void)
 /*                                                                                   */
 /* Parameters:                                                                       */
 /*   sicFeature     SIC_SET_CLOCK, SIC_SET_CALLBACK, SIC_GET_CARD_STATUS.            */
-/*   sicArg0        Depend on feature setting.                                       */
-/*   sicArg1        Depend on feature setting.                                       */
-/*   sicArg2        Depend on feature setting.                                       */
+/*   sicArg0		Depend on feature setting.										 */
+/*   sicArg1		Depend on feature setting.										 */
+/*   sicArg2		Depend on feature setting.										 */
 /*                                                                                   */
 /* Returns:                                                                          */
 /*   None.                                                                           */
@@ -105,11 +105,11 @@ void sicClose(void)
 /*                                                                                   */
 /* Description:                                                                      */
 /*   This function set the FMI engine clock. Card insert/remove callback             */
-/*  SIC_SET_CLOCK : sicArg0 used to set clock by KHz                                 */
+/*	SIC_SET_CLOCK : sicArg0 used to set clock by KHz								 */
 /*                                                                                   */
-/*  SIC_SET_CALLBACK : sicArg0 used to select card type (FMI_SD_CARD0)               */
-/*  SIC_SET_CALLBACK : sicArg1 remove function pointer                               */
-/*  SIC_SET_CALLBACK : sicArg2 insert function pointer                               */
+/*	SIC_SET_CALLBACK : sicArg0 used to select card type (FMI_SD_CARD0)				 */
+/*	SIC_SET_CALLBACK : sicArg1 remove function pointer								 */
+/*	SIC_SET_CALLBACK : sicArg2 insert function pointer								 */
 /*                                                                                   */
 /*  SIC_GET_CARD_STATUS : sicArg0 got SD0 card status. TRUE is inserted,             */
 /*                        FALSE is removed.                                          */
@@ -120,24 +120,24 @@ void sicClose(void)
 /*-----------------------------------------------------------------------------------*/
 VOID sicIoctl(INT32 sicFeature, INT32 sicArg0, INT32 sicArg1, INT32 sicArg2)
 {
-    switch(sicFeature)
-    {
-        case SIC_SET_CLOCK:
-            fmiSetFMIReferenceClock(sicArg0);
-            break;
-        case SIC_SET_CALLBACK:
-            fmiSetCallBack(sicArg0, (PVOID)sicArg1, (PVOID)sicArg2);
-            break;
-        case SIC_GET_CARD_STATUS:
+	switch(sicFeature)
+	{
+		case SIC_SET_CLOCK:
+			fmiSetFMIReferenceClock(sicArg0);
+			break;
+		case SIC_SET_CALLBACK:
+			fmiSetCallBack(sicArg0, (PVOID)sicArg1, (PVOID)sicArg2);
+			break;
+		case SIC_GET_CARD_STATUS:
             if (fmiSD_CardStatus() == FMI_NO_SD_CARD)
                 *(UINT32 *)sicArg0 = FALSE;     // card removed
             else
                 *(UINT32 *)sicArg0 = TRUE;      // card inserted
-            break;
+			break;
         case SIC_SET_CARD_DETECT:
             g_SD0_card_detect = sicArg0;
             break;
-    }
+	}
 }
 
 /*-----------------------------------------------------------------------------------*/
@@ -145,13 +145,13 @@ VOID sicIoctl(INT32 sicFeature, INT32 sicArg0, INT32 sicArg1, INT32 sicArg2)
 /*   sicSdRead                                                                       */
 /*                                                                                   */
 /* Parameters:                                                                       */
-/*   sicSdPortNo    Select SD port 0 or port 1.                                      */
-/*   sdSectorNo     Sector No. to get the data from.                                 */
-/*   sdSectorCount  Sector count of this access.                                     */
-/*   sdTargetAddr   The address which data upload to SDRAM.                          */
+/*   sicSdPortNo	Select SD port 0 or port 1.										 */
+/*   sdSectorNo		Sector No. to get the data from.								 */
+/*   sdSectorCount	Sector count of this access.									 */
+/*   sdTargetAddr	The address which data upload to SDRAM.							 */
 /*                                                                                   */
 /* Returns:                                                                          */
-/*   0                  Success                                                      */
+/*   0					Success                                                      */
 /*   FMI_TIMEOUT        Access timeout                                               */
 /*   FMI_NO_SD_CARD     Card removed                                                 */
 /*   FMI_SD_CRC7_ERROR  Command/Response error                                       */
@@ -161,38 +161,38 @@ VOID sicIoctl(INT32 sicFeature, INT32 sicArg0, INT32 sicArg1, INT32 sicArg2)
 /*   None.                                                                           */
 /*                                                                                   */
 /* Description:                                                                      */
-/*   This function used to get the data from SD card                                 */
+/*   This function used to get the data from SD card					             */
 /*-----------------------------------------------------------------------------------*/
 INT sicSdRead(INT32 sdSectorNo, INT32 sdSectorCount, INT32 sdTargetAddr)
 {
-    return sicSdRead0(sdSectorNo, sdSectorCount, sdTargetAddr);
+	return sicSdRead0(sdSectorNo, sdSectorCount, sdTargetAddr);
 }
 
 INT sicSdRead0(INT32 sdSectorNo, INT32 sdSectorCount, INT32 sdTargetAddr)
 {
-    if (fmiSD_CardSel(0))
-        return FMI_NO_SD_CARD;
+	if (fmiSD_CardSel(0))
+		return FMI_NO_SD_CARD;
 
-    outpw(REG_FMICR, FMI_SD_EN);
-    return fmiSD_Read_in(pSD0, sdSectorNo, sdSectorCount, sdTargetAddr);
+	outpw(REG_FMICR, FMI_SD_EN);
+	return fmiSD_Read_in(pSD0, sdSectorNo, sdSectorCount, sdTargetAddr);
 }
 
 INT sicSdRead1(INT32 sdSectorNo, INT32 sdSectorCount, INT32 sdTargetAddr)
 {
-    if (fmiSD_CardSel(1))
-        return FMI_NO_SD_CARD;
+	if (fmiSD_CardSel(1))
+		return FMI_NO_SD_CARD;
 
-    outpw(REG_FMICR, FMI_SD_EN);
-    return fmiSD_Read_in(pSD1, sdSectorNo, sdSectorCount, sdTargetAddr);
+	outpw(REG_FMICR, FMI_SD_EN);
+	return fmiSD_Read_in(pSD1, sdSectorNo, sdSectorCount, sdTargetAddr);
 }
 
 INT sicSdRead2(INT32 sdSectorNo, INT32 sdSectorCount, INT32 sdTargetAddr)
 {
-    if (fmiSD_CardSel(2))
-        return FMI_NO_SD_CARD;
+	if (fmiSD_CardSel(2))
+		return FMI_NO_SD_CARD;
 
-    outpw(REG_FMICR, FMI_SD_EN);
-    return fmiSD_Read_in(pSD2, sdSectorNo, sdSectorCount, sdTargetAddr);
+	outpw(REG_FMICR, FMI_SD_EN);
+	return fmiSD_Read_in(pSD2, sdSectorNo, sdSectorCount, sdTargetAddr);
 }
 
 /*-----------------------------------------------------------------------------
@@ -209,9 +209,9 @@ INT sicSdRead0_SG(INT32 sdSectorNo, INT32 sdSectorCount, INT32 sdTargetAddr, INT
 
     outpw(REG_DMACCSR, inpw(REG_DMACCSR) | SG_EN);  // enable Scatter Gather feature
     if (outOfOrder)
-        result = sicSdRead0(sdSectorNo, sdSectorCount, sdTargetAddr | PAD_ORDER);   // out of order
-    else
-        result = sicSdRead0(sdSectorNo, sdSectorCount, sdTargetAddr & ~PAD_ORDER);  // in order
+	    result = sicSdRead0(sdSectorNo, sdSectorCount, sdTargetAddr | PAD_ORDER);   // out of order
+	else
+	    result = sicSdRead0(sdSectorNo, sdSectorCount, sdTargetAddr & ~PAD_ORDER);  // in order
     outpw(REG_DMACCSR, inpw(REG_DMACCSR) & ~SG_EN); // disable Scatter Gather feature
     return result;
 }
@@ -222,9 +222,9 @@ INT sicSdRead1_SG(INT32 sdSectorNo, INT32 sdSectorCount, INT32 sdTargetAddr, INT
 
     outpw(REG_DMACCSR, inpw(REG_DMACCSR) | SG_EN);  // enable Scatter Gather feature
     if (outOfOrder)
-        result = sicSdRead1(sdSectorNo, sdSectorCount, sdTargetAddr | PAD_ORDER);   // out of order
-    else
-        result = sicSdRead1(sdSectorNo, sdSectorCount, sdTargetAddr & ~PAD_ORDER);  // in order
+	    result = sicSdRead1(sdSectorNo, sdSectorCount, sdTargetAddr | PAD_ORDER);   // out of order
+	else
+	    result = sicSdRead1(sdSectorNo, sdSectorCount, sdTargetAddr & ~PAD_ORDER);  // in order
     outpw(REG_DMACCSR, inpw(REG_DMACCSR) & ~SG_EN); // disable Scatter Gather feature
     return result;
 }
@@ -235,9 +235,9 @@ INT sicSdRead2_SG(INT32 sdSectorNo, INT32 sdSectorCount, INT32 sdTargetAddr, INT
 
     outpw(REG_DMACCSR, inpw(REG_DMACCSR) | SG_EN);  // enable Scatter Gather feature
     if (outOfOrder)
-        result = sicSdRead2(sdSectorNo, sdSectorCount, sdTargetAddr | PAD_ORDER);   // out of order
-    else
-        result = sicSdRead2(sdSectorNo, sdSectorCount, sdTargetAddr & ~PAD_ORDER);  // in order
+	    result = sicSdRead2(sdSectorNo, sdSectorCount, sdTargetAddr | PAD_ORDER);   // out of order
+	else
+	    result = sicSdRead2(sdSectorNo, sdSectorCount, sdTargetAddr & ~PAD_ORDER);  // in order
     outpw(REG_DMACCSR, inpw(REG_DMACCSR) & ~SG_EN); // disable Scatter Gather feature
     return result;
 }
@@ -248,13 +248,13 @@ INT sicSdRead2_SG(INT32 sdSectorNo, INT32 sdSectorCount, INT32 sdTargetAddr, INT
 /*   sicSdWrite                                                                      */
 /*                                                                                   */
 /* Parameters:                                                                       */
-/*   sicSdPortNo    Select SD port 0 or port 1.                                      */
-/*   sdSectorNo     Sector No. to get the data from.                                 */
-/*   sdSectorCount  Sector count of this access.                                     */
-/*   sdSourcetAddr  The address which data download data from SDRAM.                 */
+/*   sicSdPortNo	Select SD port 0 or port 1.										 */
+/*   sdSectorNo		Sector No. to get the data from.								 */
+/*   sdSectorCount	Sector count of this access.									 */
+/*   sdSourcetAddr	The address which data download data from SDRAM.				 */
 /*                                                                                   */
 /* Returns:                                                                          */
-/*   0                  Success                                                      */
+/*   0					Success                                                      */
 /*   FMI_TIMEOUT        Access timeout                                               */
 /*   FMI_NO_SD_CARD     Card removed                                                 */
 /*   FMI_SD_CRC7_ERROR  Command/Response error                                       */
@@ -264,38 +264,38 @@ INT sicSdRead2_SG(INT32 sdSectorNo, INT32 sdSectorCount, INT32 sdTargetAddr, INT
 /*   None.                                                                           */
 /*                                                                                   */
 /* Description:                                                                      */
-/*   This function used to write the data to SD card                                 */
+/*   This function used to write the data to SD card					             */
 /*-----------------------------------------------------------------------------------*/
 INT sicSdWrite(INT32 sdSectorNo, INT32 sdSectorCount, INT32 sdSourceAddr)
 {
-    return sicSdWrite0(sdSectorNo, sdSectorCount, sdSourceAddr);
+	return sicSdWrite0(sdSectorNo, sdSectorCount, sdSourceAddr);
 }
 
 INT sicSdWrite0(INT32 sdSectorNo, INT32 sdSectorCount, INT32 sdSourceAddr)
 {
-    if (fmiSD_CardSel(0))
-        return FMI_NO_SD_CARD;
+	if (fmiSD_CardSel(0))
+		return FMI_NO_SD_CARD;
 
-    outpw(REG_FMICR, FMI_SD_EN);
-    return fmiSD_Write_in(pSD0, sdSectorNo, sdSectorCount, sdSourceAddr);
+	outpw(REG_FMICR, FMI_SD_EN);
+	return fmiSD_Write_in(pSD0, sdSectorNo, sdSectorCount, sdSourceAddr);
 }
 
 INT sicSdWrite1(INT32 sdSectorNo, INT32 sdSectorCount, INT32 sdSourceAddr)
 {
-    if (fmiSD_CardSel(1))
-        return FMI_NO_SD_CARD;
+	if (fmiSD_CardSel(1))
+		return FMI_NO_SD_CARD;
 
-    outpw(REG_FMICR, FMI_SD_EN);
-    return fmiSD_Write_in(pSD1, sdSectorNo, sdSectorCount, sdSourceAddr);
+	outpw(REG_FMICR, FMI_SD_EN);
+	return fmiSD_Write_in(pSD1, sdSectorNo, sdSectorCount, sdSourceAddr);
 }
 
 INT sicSdWrite2(INT32 sdSectorNo, INT32 sdSectorCount, INT32 sdSourceAddr)
 {
-    if (fmiSD_CardSel(2))
-        return FMI_NO_SD_CARD;
+	if (fmiSD_CardSel(2))
+		return FMI_NO_SD_CARD;
 
-    outpw(REG_FMICR, FMI_SD_EN);
-    return fmiSD_Write_in(pSD2, sdSectorNo, sdSectorCount, sdSourceAddr);
+	outpw(REG_FMICR, FMI_SD_EN);
+	return fmiSD_Write_in(pSD2, sdSectorNo, sdSectorCount, sdSourceAddr);
 }
 
 
@@ -313,9 +313,9 @@ INT sicSdWrite0_SG(INT32 sdSectorNo, INT32 sdSectorCount, INT32 sdTargetAddr, IN
 
     outpw(REG_DMACCSR, inpw(REG_DMACCSR) | SG_EN);  // enable Scatter Gather feature
     if (outOfOrder)
-        result = sicSdWrite0(sdSectorNo, sdSectorCount, sdTargetAddr | PAD_ORDER);   // out of order
-    else
-        result = sicSdWrite0(sdSectorNo, sdSectorCount, sdTargetAddr & ~PAD_ORDER);  // in order
+	    result = sicSdWrite0(sdSectorNo, sdSectorCount, sdTargetAddr | PAD_ORDER);   // out of order
+	else
+	    result = sicSdWrite0(sdSectorNo, sdSectorCount, sdTargetAddr & ~PAD_ORDER);  // in order
     outpw(REG_DMACCSR, inpw(REG_DMACCSR) & ~SG_EN); // disable Scatter Gather feature
     return result;
 }
@@ -326,9 +326,9 @@ INT sicSdWrite1_SG(INT32 sdSectorNo, INT32 sdSectorCount, INT32 sdTargetAddr, IN
 
     outpw(REG_DMACCSR, inpw(REG_DMACCSR) | SG_EN);  // enable Scatter Gather feature
     if (outOfOrder)
-        result = sicSdWrite1(sdSectorNo, sdSectorCount, sdTargetAddr | PAD_ORDER);   // out of order
-    else
-        result = sicSdWrite1(sdSectorNo, sdSectorCount, sdTargetAddr & ~PAD_ORDER);  // in order
+	    result = sicSdWrite1(sdSectorNo, sdSectorCount, sdTargetAddr | PAD_ORDER);   // out of order
+	else
+	    result = sicSdWrite1(sdSectorNo, sdSectorCount, sdTargetAddr & ~PAD_ORDER);  // in order
     outpw(REG_DMACCSR, inpw(REG_DMACCSR) & ~SG_EN); // disable Scatter Gather feature
     return result;
 }
@@ -339,9 +339,9 @@ INT sicSdWrite2_SG(INT32 sdSectorNo, INT32 sdSectorCount, INT32 sdTargetAddr, IN
 
     outpw(REG_DMACCSR, inpw(REG_DMACCSR) | SG_EN);  // enable Scatter Gather feature
     if (outOfOrder)
-        result = sicSdWrite2(sdSectorNo, sdSectorCount, sdTargetAddr | PAD_ORDER);   // out of order
-    else
-        result = sicSdWrite2(sdSectorNo, sdSectorCount, sdTargetAddr & ~PAD_ORDER);  // in order
+	    result = sicSdWrite2(sdSectorNo, sdSectorCount, sdTargetAddr | PAD_ORDER);   // out of order
+	else
+	    result = sicSdWrite2(sdSectorNo, sdSectorCount, sdTargetAddr & ~PAD_ORDER);  // in order
     outpw(REG_DMACCSR, inpw(REG_DMACCSR) & ~SG_EN); // disable Scatter Gather feature
     return result;
 }
